Fixed reports listing
